requirements.txt
hachoir==3.3.0
pillow==11.3.0
